% Auto-generated: do not edit by hand
\name{reactScoreIndicator}

\alias{reactScoreIndicator}

\title{ReactScoreIndicator component}

\description{

}

\usage{
reactScoreIndicator(maxValue=NULL, stepsColors=NULL, textStyle=NULL,
toClassification=NULL, value=NULL, width=NULL)
}

\arguments{
\item{maxValue}{Numeric. }

\item{stepsColors}{List of characters. }

\item{textStyle}{Named list. }

\item{toClassification}{Logical. }

\item{value}{Numeric. }

\item{width}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

